test: test design refactoring (body) (#2921)
