+++ doc/changelog.d/2921.test.md
@@ -1,0 +1,1 @@
+Test design refactoring (body)
